requirements.txt
django 
djangorestframework 
opencv-python 
mediapipe 
numpy